feat: add update function to update system packages and tools
The `update` function has been added to the `.config/fish/config.fish` file. This function is used to update system packages and tools on different environments. It includes updating Homebrew packages on macOS, updating APT packages on Ubuntu, updating anyenv, rustup, rye, and devbox if they are installed, and updating completion files for rye and devbox.

This commit also includes some changes in the aliases section, such as commenting out unnecessary aliases and updating the `ghqpull` alias.

Additionally, some changes have been made to the PATH configuration section, such as adding `$PATH` for Rust and rye if the corresponding directories exist.

Lastly, the comment section for the "etc" section has been added to separate different sections in the file.